requirements.txt
Django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.0
Pillow==10.1.0
reportlab==4.0.7
python-decouple==3.8
django-cors-headers==4.3.1
drf-spectacular==0.27.0
django-filter==23.5
